feat: 주문서 생성 API 구현
- Request/Response DTO 작성 (하위 DTO는 재사용 없어 내부 record로 포함)
- BusinessException ErrorCode 추가
- 검증 순서: Funding 검증(존재/기간) → Reward 검증(중복/존재/차감 전 재고)
- 도메인 객체에 검증 메서드 추가 (isFundingPeriod, isStockSufficient, calculateAmount)
- 가격 타입 Long 통일 (DB: BIGINT)

+++ services/order-service/src/main/java/com/bds/order/infrastructure/reward/RewardJpaEntity.java
@@ -32,18 +32,18 @@ public class RewardJpaEntity {
     @Column(nullable = false)
     private int limitQty;
 
-    private Integer remainQty;
+    private int remainQty;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20)
     private BadgeType badgeType;
 
     @Column(nullable = false)
-    private int price;
+    private Long price;
 
     @Column(nullable = false)
     private LocalDateTime offerAt;
 
     @Column(nullable = false)
-    private int shippingCharge;
+    private Long shippingCharge;
 }
